awssqs: move adapter
